+++ src/features/flags/flags.ts
@@ -9,7 +9,6 @@ import { Logger } from "../logger";
  */
 export function setDefaultFlags(app: App) {
   setFlag(app, "disable-seccomp-filter-sandbox");
-  setFlag(app, "disable-features", "MediaSessionService");
 }
 
 /**